moved rectangle class
